refactor(agent-runtime): give an automation run its own executor
The scheduler owned everything: the tick, the due comparison, the overlap lock,
model resolution, starting a runtime session, prompting it and shaping the run
record. Splitting execution out leaves the scheduler with the one job its name
promises — when a run is due and that a second run of the same id never starts —
and puts the run itself behind runAutomation(), which takes an automation and
returns the record of what happened without touching the store.

The executor is where the target now decides how a run starts. A global target
keeps the fresh-session behaviour it always had. A thread target resumes the
conversation's own rollout, and refuses to run when that rollout is not on disk
rather than silently starting a detached conversation the automation would then
be attached to in name only. The stable automation:${id} runtime key stays, so
a resumed thread still runs under the automation's own session rather than the
one the UI may have open on the same rollout.
