+++ services/agent-runtime/src/automation-scheduler.ts
@@ -1,3 +1,4 @@
+import { runAutomation } from "./automation-runner";
 import {
   getAutomation,
   listAutomations,
@@ -5,14 +6,10 @@ import {
   patchAutomation,
   recordAutomationRun,
   type Automation,
-  type AutomationRun,
 } from "./automations-store";
 import { getGlobalSingleton } from "./instances";
-import { piRuntimeManager } from "./pi-runtime";
-import { refreshPiModels } from "./pi-runtime-models";
-import { lastAssistantResult } from "./session-text";
-import { listProjectsFromStore } from "./projects-store";
-import type { AgentModel } from "../../../shared/agent/models";
+
+export { automationRunError } from "./automation-runner";
 
 const TICK_MS = 30_000;
 
@@ -28,114 +25,17 @@ function state(): SchedulerState {
   }));
 }
 
-function runPrompt(automation: Automation): string {
-  const preamble = automation.lastRun?.summary
-    ? `Previous run summary (context, may be stale):\n${automation.lastRun.summary}\n\n---\n\n`
-    : "";
-  return `${preamble}${automation.prompt}`;
-}
-
-export function automationRunError(lastError: string | null, summary: string): string | null {
-  if (lastError) return lastError;
-  return summary.trim() ? null : "Automation completed without an assistant response.";
-}
-
-export type AutomationModelResolution =
-  | { ok: true; modelId: string; fallback: boolean }
-  | { ok: false; error: string };
-
-export const NO_ACTIVE_MODEL_ERROR =
-  "No model is loaded right now, so this automation could not run. Load a model in Local Studio and try again.";
-export const MODEL_LOOKUP_ERROR =
-  "Could not read the list of runtime models, so this automation could not pick a model to run on.";
-
-export function resolveAutomationModel(
-  requestedModelId: string,
-  models: readonly AgentModel[],
-): AutomationModelResolution {
-  const requested = models.find(
-    (model) => model.id === requestedModelId || model.rawId === requestedModelId,
-  );
-  if (requested?.active) return { ok: true, modelId: requestedModelId, fallback: false };
-  const active = models.find((model) => model.active);
-  if (!active) return { ok: false, error: NO_ACTIVE_MODEL_ERROR };
-  return { ok: true, modelId: active.id, fallback: true };
-}
-
-function failedRun(automation: Automation, error: string): AutomationRun {
-  return {
-    at: new Date().toISOString(),
-    piSessionId: null,
-    cwd: automation.cwd,
-    projectId: null,
-    outcome: "error",
-    summary: "",
-    error,
-    requestedModelId: automation.modelId,
-  };
-}
-
 export async function runAutomationNow(id: string): Promise<Automation | null> {
   const scheduler = state();
   const automation = await getAutomation(id);
   if (!automation || scheduler.running.has(id)) return null;
   scheduler.running.add(id);
-  const runtimeSessionId = `automation:${id}`;
-  const scheduleNext = () => nextRunAt(automation.schedule, new Date()).toISOString();
   try {
-    let models: readonly AgentModel[];
-    try {
-      ({ models } = await refreshPiModels());
-    } catch (error) {
-      const detail = error instanceof Error ? ` ${error.message}` : "";
-      return await recordAutomationRun(
-        id,
-        failedRun(automation, `${MODEL_LOOKUP_ERROR}${detail}`),
-        scheduleNext(),
-      );
-    }
-    const resolution = resolveAutomationModel(automation.modelId, models);
-    if (!resolution.ok) {
-      return await recordAutomationRun(id, failedRun(automation, resolution.error), scheduleNext());
-    }
-    const modelFields = resolution.fallback
-      ? {
-          requestedModelId: automation.modelId,
-          actualModelId: resolution.modelId,
-          fallbackReason: "requested_model_inactive" as const,
-        }
-      : { requestedModelId: automation.modelId, actualModelId: resolution.modelId };
-    const { session } = piRuntimeManager.getSessionForLookup(runtimeSessionId, null);
-    await session.ensureStarted(resolution.modelId, automation.cwd || undefined, null, {});
-    await session.prompt(runPrompt(automation), () => {});
-    const status = session.status;
-    const piSessionId = status.piSessionId;
-    const result = piSessionId
-      ? lastAssistantResult(status.cwd, piSessionId)
-      : { text: "", error: null };
-    const error = automationRunError(status.lastError ?? result.error, result.text);
-    const projectId =
-      listProjectsFromStore().find((project) => project.path === status.cwd)?.id ?? null;
-    void session.stop().catch(() => undefined);
+    const run = await runAutomation(automation);
     return await recordAutomationRun(
       id,
-      {
-        at: new Date().toISOString(),
-        piSessionId,
-        cwd: status.cwd,
-        projectId,
-        outcome: error ? "error" : "ok",
-        summary: result.text,
-        ...(error ? { error } : {}),
-        ...modelFields,
-      },
-      scheduleNext(),
-    );
-  } catch (error) {
-    return await recordAutomationRun(
-      id,
-      failedRun(automation, error instanceof Error ? error.message : "Automation run failed"),
-      scheduleNext(),
+      run,
+      nextRunAt(automation.schedule, new Date()).toISOString(),
     );
   } finally {
     scheduler.running.delete(id);
@@ -159,7 +59,7 @@ async function tick(): Promise<void> {
       continue;
     }
     if (new Date(automation.nextRunAt) <= now) {
-      void runAutomationNow(automation.id);
+      void runAutomationNow(automation.id).catch(() => undefined);
     }
   }
 }
